Zagwire: sample pulses in interrupt routine, evaluate in non-irq context
The negative pulse times of the 20 bits of the zagwire packet are recorded in the interrupt,
but evaluated later in the application. This leads to improved error detection. During testing we had 0 gliches in ~30mins.
It is not possible to reduce the error rate to zero due to the high interrupt activity in the bus-module, which disturbs the input capture timer interrupt.
If the error rate is considered too high later, a different sensor with an "AVR-friendly" and KWB friendly
protocol needs to be selected.