pushbutton8 sw: test version for two pushbutton8 modules
Pushing one button of the left side sends message to other module and the LED next to the corresponding button lights up or blinks.
Pushing one button of the right side resets corresponding LED on the other module.
Module addresses of sender and receiver must be set in "application.c" line 32 and 33.